+++ moa/src/main/java/moa/classifiers/meta/imbalanced/OnlineAdaC2.java
@@ -1,5 +1,5 @@
 /*
- *  OnlineUnderOverBagging.java
+ *  OnlineAdaC2.java
  *  
  *  @author Alessio Bernardo (alessio dot bernardo at polimi dot dot it)
  *
@@ -16,9 +16,11 @@
  *  limitations under the License.
         
  */
-package moa.classifiers.meta;
+package moa.classifiers.meta.imbalanced;
 
 import com.yahoo.labs.samoa.instances.Instance;
+
+
 import moa.capabilities.CapabilitiesHandler;
 import moa.capabilities.Capability;
 import moa.capabilities.ImmutableCapabilities;
@@ -29,6 +31,7 @@ import moa.core.DoubleVector;
 import moa.core.Measurement;
 import moa.core.Utils;
 import moa.options.ClassOption;
+import com.github.javacliparser.FloatOption;
 import com.github.javacliparser.FlagOption;
 import com.github.javacliparser.IntOption;
 import java.util.ArrayList;
@@ -36,18 +39,14 @@ import java.util.Random;
 import moa.classifiers.core.driftdetection.ADWIN;
 
 
-
 /**
- *  Online UnderOverBagging is the online version of the ensemble method.
+ *  OnlineAdaC2 is the adaptation of the ensemble learner to data streams
  *
- * <p>In case of imbalanced classes UnderOverBagging uses the strategy of under-sampling
-     the majority class and oversampling the minority class. In addition the sampling
-     rate can be also varied over the bagging iterations, which further boosts the
-     diversity of the base learners. <p>
-
-   <p>The derivation of the online UnderOverBagging algorithm is made through the observation
-     that a Binomial distribution with sampling rate :math:`\frac{C}{N}` corresponds to a
-     poisson distribution with :math:`\lambda=C`.</p>
+ * <p>AdaC2 is a boosting algorithm that takes the different misclassification costs
+    into consideration when calculating the classier weights, and updates the sample
+    weight. AdaC2 increases more the weights on the misclassified positive samples,
+    compared to misclassified negative samples.:math:`C_P` and :math:`C_N` are the
+    positive and negative costs.</p>
 
     <p>This online ensemble learner method is improved by the addition of an ADWIN change
     detector. ADWIN stands for Adaptive Windowing. It works by keeping updated
@@ -61,7 +60,8 @@ import moa.classifiers.core.driftdetection.ADWIN;
  * <p>Parameters:</p> <ul>
  * <li>-l : Each classiﬁer to train of the ensemble is an instance of the base estimator.</li>
  * <li>-s : The size of the ensemble, in other words, how many classifiers to train.</li>
- * <li>-i : The sampling rate of the positive instances.</li>
+ * <li>-p : The cost of misclassifying a positive sample.</li>
+ * <li>-n : The cost of misclassifying a negative sample.</li>
  * <li>-d : Should use ADWIN as drift detector? If enabled it is used by the method 
  * 	to track the performance of the classifiers and adapt when a drift is detected.</li>
  * <li>-r : Seed for the random state.</li>
@@ -70,7 +70,7 @@ import moa.classifiers.core.driftdetection.ADWIN;
  * @author Alessio Bernardo (alessio dot bernardo at polimi dot dot it)
  * @version $Revision: 1 $
  */
-public class OnlineUnderOverBagging extends AbstractClassifier implements MultiClassClassifier,
+public class OnlineAdaC2 extends AbstractClassifier implements MultiClassClassifier,
                                                                         CapabilitiesHandler {
 
     @Override
@@ -86,8 +86,11 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
     public IntOption ensembleSizeOption = new IntOption("ensembleSize", 's',
         "The size of the ensemble.", 10, 1, Integer.MAX_VALUE);        
     
-    public IntOption samplingRateOption = new IntOption("samplingRate", 'i',
-            "The sampling rate of the positive instances.", 2, 1, 10);
+    public FloatOption costPositiveOption = new FloatOption("costPositive", 'p',
+        "The cost of misclassifying a positive sample.", 1, 0.1, 1);
+    
+    public FloatOption costNegativeOption = new FloatOption("costNegative", 'n',
+            "The cost of misclassifying a negative sample.", 0.1, 0.1, 1);
     
     public FlagOption disableDriftDetectionOption = new FlagOption("disableDriftDetection", 'd',
             "Should use ADWIN as drift detector?");
@@ -97,11 +100,19 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
     
     protected Classifier baseLearner;
     protected int nEstimators;    
-    protected int samplingRate; 
+    protected double costPositive;
+    protected double costNegative;
     protected boolean driftDetection;        
     protected ArrayList<Classifier> ensemble = new ArrayList<Classifier>();
     protected Random randomState;
-    protected ArrayList<ADWIN> adwinEnsemble = new ArrayList<ADWIN>();      
+    protected ArrayList<ADWIN> adwinEnsemble = new ArrayList<ADWIN>();
+    protected ArrayList<Double> lambdaTP = new ArrayList<Double>();
+    protected ArrayList<Double> lambdaTN = new ArrayList<Double>();
+    protected ArrayList<Double> lambdaFP = new ArrayList<Double>();
+    protected ArrayList<Double> lambdaFN = new ArrayList<Double>();
+    protected ArrayList<Double> lambdaSum = new ArrayList<Double>();
+    protected ArrayList<Double> wAcc = new ArrayList<Double>();
+    protected ArrayList<Double> wErr = new ArrayList<Double>();     
     
     @Override
     public void resetLearningImpl() {
@@ -109,13 +120,21 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
     	this.baseLearner = (Classifier) getPreparedClassOption(this.baseLearnerOption);
     	this.baseLearner.resetLearning();
         this.nEstimators = this.ensembleSizeOption.getValue();        
-        this.samplingRate = this.samplingRateOption.getValue();
+        this.costPositive = this.costPositiveOption.getValue();
+        this.costNegative = this.costNegativeOption.getValue();
         this.driftDetection = !this.disableDriftDetectionOption.isSet();                
         for (int i = 0; i < this.nEstimators; i++) {
         	this.ensemble.add(this.baseLearner.copy());         	        
         	if (this.driftDetection) {
         		this.adwinEnsemble.add(new ADWIN());
-        	}        	
+        	}
+        	this.lambdaTP.add(0.0);
+            this.lambdaTN.add(0.0);
+            this.lambdaFP.add(0.0);
+            this.lambdaFN.add(0.0);
+            this.lambdaSum.add(0.0);
+            this.wAcc.add(0.0);
+            this.wErr.add(0.0);   
 		}
         this.randomState = new Random(this.seedOption.getValue());           
     }
@@ -126,23 +145,41 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
         	resetLearningImpl();
         }  
         adjustEnsembleSize(instance.numClasses());
-                
+        
+        double lambda = 1.0;
         boolean changeDetected = false;        
-        double lambda = 0.0;
         
         for (int i = 0 ; i < this.ensemble.size(); i++) {
-        	double a = (double)(i + 1) / (double)this.nEstimators;
-        	if (instance.classValue() == 1.0) {
-        		lambda = a * this.samplingRate;
-        	}
-        	else {
-        		lambda = a;
-        	}        	        	
+			this.lambdaSum.set(i, this.lambdaSum.get(i) + lambda);			
 			double k = getPoisson(lambda);
 			if (k > 0) {
 				for (int b = 0; b < k; b++) {
 					this.ensemble.get(i).trainOnInstance(instance);					
-				}	
+				}
+				if (Utils.maxIndex(this.ensemble.get(i).getVotesForInstance(instance)) == 1.0 && instance.classValue() == 1.0) {
+					this.lambdaTP.set(i, this.lambdaTP.get(i) + (this.costPositive * lambda));
+					this.wAcc.set(i, (this.lambdaTP.get(i) + this.lambdaTN.get(i)) / this.lambdaSum.get(i));
+					this.wErr.set(i, (this.lambdaFP.get(i) + this.lambdaFN.get(i)) / this.lambdaSum.get(i));
+					lambda = (this.costPositive * lambda) / (2 * this.wAcc.get(i));
+				}
+				else if (Utils.maxIndex(this.ensemble.get(i).getVotesForInstance(instance)) == 0.0 && instance.classValue() == 0.0) {
+					this.lambdaTN.set(i, this.lambdaTN.get(i) + (this.costNegative * lambda));
+					this.wAcc.set(i, (this.lambdaTP.get(i) + this.lambdaTN.get(i)) / this.lambdaSum.get(i));
+					this.wErr.set(i, (this.lambdaFP.get(i) + this.lambdaFN.get(i)) / this.lambdaSum.get(i));
+					lambda = (this.costNegative * lambda) / (2 * this.wAcc.get(i));
+				}
+				else if (Utils.maxIndex(this.ensemble.get(i).getVotesForInstance(instance)) == 0.0 && instance.classValue() == 1.0) {
+					this.lambdaFN.set(i, this.lambdaFN.get(i) + (this.costPositive * lambda));
+					this.wAcc.set(i, (this.lambdaTP.get(i) + this.lambdaTN.get(i)) / this.lambdaSum.get(i));
+					this.wErr.set(i, (this.lambdaFP.get(i) + this.lambdaFN.get(i)) / this.lambdaSum.get(i));
+					lambda = (this.costPositive * lambda) / (2 * this.wErr.get(i));
+				}
+				else if (Utils.maxIndex(this.ensemble.get(i).getVotesForInstance(instance)) == 1.0 && instance.classValue() == 0.0) {
+					this.lambdaFP.set(i, this.lambdaFP.get(i) + (this.costNegative * lambda));
+					this.wAcc.set(i, (this.lambdaTP.get(i) + this.lambdaTN.get(i)) / this.lambdaSum.get(i));
+					this.wErr.set(i, (this.lambdaFP.get(i) + this.lambdaFN.get(i)) / this.lambdaSum.get(i));
+					lambda = (this.costNegative * lambda) / (2 * this.wErr.get(i));
+				}
 			}
 			if (this.driftDetection) {
 				double pred = Utils.maxIndex(this.ensemble.get(i).getVotesForInstance(instance));
@@ -180,7 +217,10 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
 
         for(int i = 0 ; i < this.ensemble.size() ; ++i) {
             DoubleVector vote = new DoubleVector(this.ensemble.get(i).getVotesForInstance(testInstance));
-            if (vote.sumOfValues() > 0.0) {                                                                                                                              
+            if (vote.sumOfValues() > 0.0) {                                                                                  
+                for(int v = 0 ; v < vote.numValues() ; ++v) {
+                    vote.setValue(v, vote.getValue(v) * Math.log(this.wAcc.get(i) / this.wErr.get(i)));
+                }                                
             	vote.normalize();
                 combinedVote.addValues(vote);                
             }
@@ -207,7 +247,14 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
     		for (int i = this.nEstimators; i < nClasses; i++) {
     			this.ensemble.add(this.baseLearner.copy()); 
     			this.nEstimators ++;
-    			this.adwinEnsemble.add(new ADWIN());    			
+    			this.adwinEnsemble.add(new ADWIN());
+    			this.lambdaTP.add(0.0);
+                this.lambdaTN.add(0.0);
+                this.lambdaFP.add(0.0);
+                this.lambdaFN.add(0.0);
+                this.lambdaSum.add(0.0);
+                this.wAcc.add(0.0);
+                this.wErr.add(0.0);  
 			}
     	}
     }
@@ -226,7 +273,7 @@ public class OnlineUnderOverBagging extends AbstractClassifier implements MultiC
 
     @Override
     public ImmutableCapabilities defineImmutableCapabilities() {
-        if (this.getClass() == OnlineUnderOverBagging.class)
+        if (this.getClass() == OnlineAdaC2.class)
             return new ImmutableCapabilities(Capability.VIEW_STANDARD, Capability.VIEW_LITE);
         else
             return new ImmutableCapabilities(Capability.VIEW_STANDARD);
